requirements.txt
numpy~=1.26.0
torch~=2.2.1
pandas~=2.1.0
torchvision~=0.17.1
matplotlib~=3.8.0
PyYAML~=5.4.1
scikit-learn~=1.3.0